backend/pyproject.toml
[project]
name = "tz-fabric"
version = "1.0.1"
description = "This is fabric analyzer backend"
authors = [{ name = "recursivezero", email = "recursivezero@outlook.com" }]
readme = "README.md"
license = { file = "LICENSE-PYTHON" }
keywords = ["fabric", "analyzer", "fastapi", "gpt", "genai"]
classifiers = [
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Framework :: FastAPI",
]
requires-python = ">=3.11,<3.13"

# Core, lightweight runtime deps only
dependencies = [
  "fastapi>=0.116.1,<0.117.0",
  "dotenv>=0.9.9,<0.10.0",
  "google-generativeai>=0.8.5,<0.9.0",
  "pillow>=11.3.0,<12.0.0",
  "python-multipart>=0.0.20,<0.1.0",
  "jinja2>=3.1.6,<4.0.0",
  "pymongo>=4.13.2,<5.0.0",
  "numpy (>=2,<2.3.0)",
  "chromadb>=1.0.17,<2.0.0",
  "pydantic>=2.11.7,<3.0.0",
  "pydantic-settings>=2.10.1,<3.0.0",
  "modelcontextprotocol>=0.1.0,<0.2.0",
  "langgraph>=0.6.6,<1.0.0",
  "langchain>=0.3.27,<1.0.0",
  "langchain-core>=0.3.75,<1.0.0",
  "langchain-xai>=0.2.5,<1.0.0",
  "langchain-groq>=0.3.7,<1.0.0",
  "mcp>=1.13.1,<2.0.0",
  "fastmcp>=2.12.5,<3.0.0",
  "click>=8.3.1,<9.0.0",
  "opencv-python (>=4.12.0.88,<5.0.0.0)",
  "scikit-image (>=0.25.2,<0.26.0)",
  "scikit-learn (>=1.8.0,<2.0.0)",
  "ultralytics (>=8.3.235,<9.0.0)"
]


[project.optional-dependencies]
# ML-only extras
ml = [
  "google-generativeai>=0.8.5,<0.9.0",
]

# Agent / LangChain stack
agent = [
  "langgraph>=0.6.6,<0.7.0",
  "langchain>=0.3.27,<0.4.0",
  "langchain-core>=0.3.75,<0.4.0",
  "langchain-xai>=0.2.5,<0.3.0",
  "langchain-groq>=0.3.7,<0.4.0",
]

# MCP integration
mcp = [
  "fastmcp>=2.12.5,<3.0.0",
  "modelcontextprotocol>=0.1.0,<0.2.0",
]

# "Heavy" bundle: vector DB, image, Mongo, NumPy + all other extras.
# Install with: pip install "tz-fabric[extra]"
extra = [
  # core heavy deps
  "pillow>=11.3.0,<12.0.0",
  "pymongo>=4.13.2,<5.0.0",
  "chromadb>=1.0.17,<2.0.0",

  # ml
  "google-generativeai>=0.8.5,<0.9.0",

  # agent
  "langgraph>=0.6.6,<0.7.0",
  "langchain>=0.3.27,<0.4.0",
  "langchain-core>=0.3.75,<0.4.0",
  "langchain-xai>=0.2.5,<0.3.0",
  "langchain-groq>=0.3.7,<0.4.0",

  # mcp
  "fastmcp>=2.12.5,<3.0.0",
  "modelcontextprotocol>=0.1.0,<0.2.0",
]


[project.urls]
homepage = "https://theadzip.com"
repository = "https://github.com/recursivezero/tz-fabric"
documentation = "https://github.com/recursivezero/tz-script/docs"

[project.scripts]
tzfabric = "cli:main"
lint = "lint:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "*"
from = "."


[tool.poetry.requires-plugins]
poetry-plugin-export = "^1.9"

[tool.poetry-auto-export]
output = "requirements.txt"
without_hashes = true
without = ["dev"]

[tool.poetry.group.dev.dependencies]
uvicorn = { extras = ["standard"], version = "^0.35.0" }
black = ">=24.10.0,<25.0.0"
flake8 = "^7.1.1"
mypy = "^1.14.1"
pip-audit = "^2.9.0"
pipreqs = "0.5.0"
ruff = "^0.13.1"
isort = "^7.0.0"
autoflake = "^2.3.1"

[tool.flake8]
max_line_length = 190
